requirements.txt
# LangChain Core
langchain
langchain-core
langchain-community

# OpenAI Integration
langchain-openai
openai

# Anthropic Integration
langchain-anthropic

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub
torch

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

#data
yfinance
pypdf

#pinecone
pinecone